clean up and move models to outer folder
