--- src/main/java/org/aerogear/plugin/intellij/mobile/ui/MobileToolWindow.form
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1">
-  <grid id="27dc6" binding="mobileToolWindowContent" layout-manager="BorderLayout" hgap="0" vgap="0">
-    <constraints>
-      <xy x="20" y="20" width="535" height="659"/>
-    </constraints>
-    <properties/>
-    <border type="none"/>
-    <children/>
-  </grid>
-</form>
