+++ src/main/java/org/aerogear/plugin/intellij/mobile/ui/MobileToolWindowFactory.java
@@ -6,13 +6,12 @@ import com.intellij.openapi.wm.*;
 import com.intellij.ui.components.JBPanel;
 import com.intellij.ui.components.JBScrollPane;
 import com.intellij.ui.content.*;
-import org.aerogear.plugin.intellij.mobile.api.Client;
-import org.aerogear.plugin.intellij.mobile.api.models.MobileServices;
-import org.aerogear.plugin.intellij.mobile.api.models.ServiceClass;
+import org.aerogear.plugin.intellij.mobile.api.MobileAPI;
+import org.aerogear.plugin.intellij.mobile.models.MobileServices;
+import org.aerogear.plugin.intellij.mobile.models.ServiceClass;
 
 import javax.swing.*;
 import java.awt.*;
-import java.io.IOException;
 
 public class MobileToolWindowFactory implements ToolWindowFactory {
 
@@ -41,14 +40,8 @@ public class MobileToolWindowFactory implements ToolWindowFactory {
         serviceClassListContainer.setLayout(new GridLayout(0, 1));
         JScrollPane scrollPane = new JBScrollPane(serviceClassListContainer);
 
-        Client cb = new Client();
-        MobileServices serviceList = null;
-        try {
-            serviceList = cb.getServices();
-        } catch (IOException e1) {
-            e1.printStackTrace();
-        }
-
+        MobileAPI mobileAPI = new MobileAPI();
+        MobileServices serviceList = mobileAPI.getServices();
         ServiceClass[] services = serviceList.getItems();
 
         for(int i = 0; i < services.length; i++) {
